+++ fdm-app/app/components/blocks/farm.tsx
@@ -1,5 +1,8 @@
 import { useState } from "react";
-import { useNavigation, Form } from "@remix-run/react";
+import { Form } from "@remix-run/react";
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useRemixForm, RemixFormProvider } from "remix-hook-form"
+import { z } from "zod"
 
 import { Button } from "@/components/ui/button"
 import {
@@ -12,8 +15,17 @@ import {
 } from "@/components/ui/card"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
+import {
+    FormControl,
+    FormDescription,
+    FormField,
+    FormItem,
+    FormLabel,
+    FormMessage,
+} from "@/components/ui/form"
 
 import { MultiSelect } from "@/components/custom/multi-select"
+import { LoadingSpinner } from "../custom/loadingspinner";
 
 
 export interface fertilizersListType {
@@ -28,71 +40,102 @@ export interface farmType {
     organicFertilizersList: fertilizersListType[]
     mineralFertilizersList: fertilizersListType[]
     action: "/app/addfarm/new"
+    FormSchema: z.Schema<any>
 }
 
+/**
+ * Renders the Farm component, which displays the farm form with validation.
+ * @param props - Properties of type `farmType`, including form data and validation schema.
+ * @returns The JSX element representing the farm form.
+ */
 export function Farm(props: farmType) {
     const organicFertilizersList = props.organicFertilizersList
     const mineralFertilizersList = props.mineralFertilizersList
+    const FormSchema = props.FormSchema
     const [selectedOrganicFertilizers, setOrganicFertilizers] = useState<string[]>(props.b_fertilizers_organic);
     const [selectedMineralFertilizers, setMineralFertilizers] = useState<string[]>(props.b_fertilizers_mineral);
 
-    const navigation = useNavigation();
+    const form = useRemixForm<z.infer<typeof FormSchema>>({
+        mode: "onTouched",
+        resolver: zodResolver(FormSchema),
+        defaultValues: {
+            b_name_farm: props.b_name_farm ?? "",
+        },
+    })
 
     return (
         <div className="flex h-screen items-center justify-center">
             <Card className="w-[350px]">
-                <Form method="post" action={props.action}>
-                    <fieldset
-                        disabled={navigation.state === "submitting"}
-                    >
-                        <CardHeader>
-                            <CardTitle>Bedrijf</CardTitle>
-                            <CardDescription>Wat voor soort bedrijf heb je?</CardDescription>
-                        </CardHeader>
-                        <CardContent>
-                            <div className="grid w-full items-center gap-4">
-                                <div className="flex flex-col space-y-1.5">
-                                    <Label htmlFor="b_name_farm">Bedrijfsnaam <span className="text-red-500">*</span></Label>
-                                    <Input id="b_name_farm" name="b_name_farm" placeholder="Bv. Jansen V.O.F." aria-required="true"/>
+                <RemixFormProvider {...form}>
+                    <Form id="formFarm" onSubmit={form.handleSubmit} method="POST">
+                        <fieldset
+                            disabled={form.formState.isSubmitting}
+                        >
+                            <CardHeader>
+                                <CardTitle>Bedrijf</CardTitle>
+                                <CardDescription>Wat voor soort bedrijf heb je?</CardDescription>
+                            </CardHeader>
+                            <CardContent>
+                                <div className="grid w-full items-center gap-4">
+                                    <div className="flex flex-col space-y-1.5">
+                                        <FormField
+                                            control={form.control}
+                                            name="b_name_farm"
+                                            render={({ field }) => (
+                                                <FormItem>
+                                                    <FormLabel>Bedrijfsnaam <span className="text-red-500">*</span></FormLabel>
+                                                    <FormControl>
+                                                        <Input placeholder="Bv. Jansen V.O.F." aria-required="true" {...field} />
+                                                    </FormControl>
+                                                    <FormDescription />
+                                                    <FormMessage />
+                                                </FormItem>
+
+                                            )}
+                                        />
+                                    </div>
+                                    <div className="flex flex-col space-y-1.5">
+                                        <Label htmlFor="b_fertilizers_organic">Welke <i>organische</i> meststoffen gebruikt u?</Label>
+                                        <MultiSelect
+                                            options={organicFertilizersList}
+                                            onValueChange={setOrganicFertilizers}
+                                            defaultValue={selectedOrganicFertilizers}
+                                            placeholder="Selecteer meststoffen"
+                                            variant="inverted"
+                                            animation={0}
+                                            maxCount={10}
+                                            name="b_fertilizers_organic"
+                                        />
+                                    </div>
+                                    <div className="flex flex-col space-y-1.5">
+                                        <Label htmlFor="b_fertilizers_mineral">Welke <i>minerale</i> meststoffen gebruikt u?</Label>
+                                        <MultiSelect
+                                            options={mineralFertilizersList}
+                                            onValueChange={setMineralFertilizers}
+                                            defaultValue={selectedMineralFertilizers}
+                                            placeholder="Selecteer meststoffen"
+                                            variant="inverted"
+                                            animation={0}
+                                            maxCount={10}
+                                            name="b_fertilizers_mineral"
+                                        />
+                                    </div>
                                 </div>
-                                <div className="flex flex-col space-y-1.5">
-                                    <Label htmlFor="b_fertilizers_organic">Welke <i>organische</i> meststoffen gebruikt u?</Label>
-                                    <MultiSelect
-                                        options={organicFertilizersList}
-                                        onValueChange={setOrganicFertilizers}
-                                        defaultValue={selectedOrganicFertilizers}
-                                        placeholder="Selecteer meststoffen"
-                                        variant="inverted"
-                                        animation={0}
-                                        maxCount={10}
-                                        name="b_fertilizers_organic"
-                                    />
-                                </div>
-                                <div className="flex flex-col space-y-1.5">
-                                    <Label htmlFor="b_fertilizers_mineral">Welke <i>minerale</i> meststoffen gebruikt u?</Label>
-                                    <MultiSelect
-                                        options={mineralFertilizersList}
-                                        onValueChange={setMineralFertilizers}
-                                        defaultValue={selectedMineralFertilizers}
-                                        placeholder="Selecteer meststoffen"
-                                        variant="inverted"
-                                        animation={0}
-                                        maxCount={10}
-                                        name="b_fertilizers_mineral"
-                                    />
-                                </div>
-                            </div>
-                        </CardContent>
-                        <CardFooter className="flex justify-between">
-                            <Button variant="outline">Terug</Button>
-                            <Button type="submit">
-                                {navigation.state === "submitting"
-                                    ? "Opslaan..."
-                                    : "Verder"}
-                            </Button>
-                        </CardFooter>
-                    </fieldset>
-                </Form>
+                            </CardContent>
+                            <CardFooter className="flex justify-between">
+                                <Button variant="outline" onClick={() => window.history.back()}>Terug</Button>
+                                <Button type="submit">
+                                    {form.formState.isSubmitting
+                                        ? <div className="flex items-center space-x-2">
+                                            <LoadingSpinner />
+                                            <span>Opslaan...</span>
+                                        </div>
+                                        : "Volgende"}
+                                </Button>
+                            </CardFooter>
+                        </fieldset>
+                    </Form>
+                </RemixFormProvider>
             </Card>
         </div>