+++ fdm-app/app/routes/signup.tsx
@@ -4,18 +4,51 @@ import type {
 } from "@remix-run/node"
 import { json, redirect } from "@remix-run/node"
 import { Form } from "@remix-run/react"
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useRemixForm, RemixFormProvider } from "remix-hook-form"
+import { z } from "zod"
 import { signUpUser, getUserFromSession } from "@svenvw/fdm-core"
 
 // Components
 import { Button } from "@/components/ui/button"
 import { Checkbox } from "@/components/ui/checkbox"
 import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
+import {
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form"
 
 // Services
 import { getSession, commitSession, destroySession } from "@/services/session.server";
 import { fdm } from "../services/fdm.server";
 import { Check } from "lucide-react"
+import { LoadingSpinner } from "@/components/custom/loadingspinner"
+import { extractFormValuesFromRequest } from "@/lib/form"
+
+const FormSchema = z.object({
+  firstname: z.coerce.string({
+    required_error: "Voornaam is verplicht",
+  }).min(2, {
+    message: "Voornaam moet minimaal 2 karakters bevatten"
+  }),
+  surname: z.coerce.string({
+    required_error: "Achternaam is verplicht",
+  }).min(2, {
+    message: "Achternaam moet minimaal 2 karakters bevatten",
+  }),
+  email: z.coerce.string({
+    required_error: "E-mail is verplicht",
+  }).email({
+    message: "Voer een geldig emailadres in",
+  }),
+  agreed: z.coerce.boolean().default(false).refine((val) => val === true, {
+    message: "Je moet akkoord gaan met de Algemene Voorwaarden en Privacyverklaring"
+  })
+})
 
 export async function loader({
   request,
@@ -52,7 +85,22 @@ export async function loader({
   });
 }
 
+/**
+ * SignUp component renders the signup form with client-side validation.
+ * Utilizes react-hook-form and Zod for form handling and validation.
+ * @returns The JSX element representing the signup form.
+ */
 export default function SignUp() {
+
+  const form = useRemixForm<z.infer<typeof FormSchema>>({
+    mode: "onTouched",
+    resolver: zodResolver(FormSchema),
+    defaultValues: {
+      agreed: false,
+    },
+  })
+
+
   return (
     <div className="w-full lg:grid lg:min-h-[600px] lg:grid-cols-2 xl:min-h-[800px]">
       <div className="flex h-screen items-center justify-center py-12">
@@ -115,59 +163,99 @@ export default function SignUp() {
               </div>
             </div>
           </div>
-          <Form method="post">
-            <div className="grid gap-4">
-              <div className="grid gap-2">
-                <Label htmlFor="firstname">Voornaam</Label>
-                <Input
-                  id="firstname"
-                  name="firstname"
-                  type="text"
-                  placeholder=""
-                  required
-                />
-              </div>
-              <div className="grid gap-2">
-                <Label htmlFor="surname">Achternaam</Label>
-                <Input
-                  id="surname"
-                  name="surname"
-                  type="text"
-                  placeholder=""
-                  required
-                />
-              </div>
-              <div className="grid gap-2">
-                <Label htmlFor="email">Email</Label>
-                <Input
-                  id="email"
-                  name="email"
-                  type="email"
-                  placeholder=""
-                  required
-                />
-              </div>
-              <div className="grid gap-2">
-                <div className="flex items-center space-x-2">
-                  <Checkbox id="terms" />
-                  <label
-                    htmlFor="terms"
-                    className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-                  >
-                    Ik ga akkoord met de Algemene Voorwaarden en Privacyverklaring
-                  </label>
+          <RemixFormProvider {...form}>
+            <Form id="formFarm" onSubmit={form.handleSubmit} method="post">
+              <fieldset
+                disabled={form.formState.isSubmitting}
+              >
+                <div className="grid gap-4">
+                  <div className="grid gap-2">
+                    <FormField
+                      control={form.control}
+                      name="firstname"
+                      render={({ field }) => (                        
+                        <FormItem>
+                          <FormLabel>Voornaam</FormLabel>
+                          <FormControl>
+                            <Input {...field} type="text" required />
+                          </FormControl>
+                          <FormDescription />
+                          <FormMessage />
+                        </FormItem>                  
+                      )}
+                    />
+                  </div>
+                  <div className="grid gap-2">
+                    <FormField
+                      control={form.control}
+                      name="surname"
+                      render={({ field }) => (
+                        <FormItem>
+                          <FormLabel>Achternaam</FormLabel>
+                          <FormControl>
+                            <Input {...field} type="text" required />
+                          </FormControl>
+                          <FormDescription />
+                          <FormMessage />
+                        </FormItem>
+                      )}
+                    />
+                  </div>
+                  <div className="grid gap-2">
+                    <FormField
+                      control={form.control}
+                      name="email"
+                      render={({ field }) => (
+                        <FormItem>
+                          <FormLabel>Email</FormLabel>
+                          <FormControl>
+                            <Input {...field} type="email" required />
+                          </FormControl>
+                          <FormDescription />
+                          <FormMessage />
+                        </FormItem>
+                      )}
+                    />
+                  </div>
+                  <div className="grid gap-2">
+                    <FormField
+                      control={form.control}
+                      name="agreed"
+                      render={({ field }) => (
+                        <FormItem className="flex items-center space-x-2">
+                          <FormControl>
+                            <Checkbox
+                              checked={field.value}
+                              onCheckedChange={field.onChange}
+                            />
+                          </FormControl>
+                          <div className="space-y-0 leading-none">
+                            <FormLabel>
+                              Ik ga akkoord met de Algemene Voorwaarden en Privacyverklaring
+                            </FormLabel>
+                            {/* <FormMessage className="text-sm font-light" />                                      */}
+                          </div>
+                        </FormItem>
+                      )}
+                    />
+                  </div>
+                  <Button type="submit">
+                    {form.formState.isSubmitting
+                      ? <div className="flex items-center space-x-2">
+                        <LoadingSpinner />
+                        <span>Registreren...</span>
+                      </div>
+                      : "Registreren"}
+                  </Button>
+                  <div className="mt-4 text-center text-sm">
+                    Wil je eerst meer weten over FDM? Kijk dan bij onze <a href="#" className="underline">
+                      Veelgestelde Vragen
+                    </a>
+                  </div>
                 </div>
-              </div>
-              <Button type="submit" className="">
-                Registreren
-              </Button>
-              <div className="mt-4 text-center text-sm">
-                Wil je eerst meer weten over FDM? Kijk dan bij onze <a href="#" className="underline">
-                  Veelgestelde Vragen
-                </a>
-              </div>
-            </div>
-          </Form>
+              </fieldset>
+            </Form>
+          </RemixFormProvider>
         </div>
       </div>
       <div className="hidden bg-muted lg:block">
@@ -181,6 +269,12 @@ export default function SignUp() {
   )
 }
 
+/**
+ * Action function for processing the signup form submission.
+ * Validates user input and creates a new user account.
+ * @param request - The incoming request object containing form data.
+ * @returns A redirect response upon successful signup or error handling.
+ */
 export async function action({
   request,
 }: ActionFunctionArgs) {
@@ -189,10 +283,8 @@ export async function action({
     request.headers.get("Cookie")
   );
 
-  const form = await request.formData()
-  const firstname = String(form.get("firstname"))
-  const surname = String(form.get("surname"))
-  const email = String(form.get("email"))
+  const formValues = await extractFormValuesFromRequest(request, FormSchema)
+  const { firstname, surname, email } = formValues;
 
   // sign up user
   const session_id = await signUpUser(fdm, firstname, surname, email)